Backend/requirements.txt
google-generativeai==0.8.4
Flask==3.1.0
flask-cors==5.0.1
python-docx==1.1.2
PyPDF2==3.0.1
pyttsx3==2.98
configparser